Rolling upgrades for Ceph Monitor Cluster
This change adds functionality to allow the ceph monitor cluster to
upgrade in a serial rolled fashion.  This will use the ceph monitor
cluster itself as a locking mechanism and only allows 1 ceph monitor
at a time to upgrade.  If a monitor has been waiting on the previous
server for more than 10 minutes and hasn't seen it finish it will
assume it died during the upgrade and proceed with its own upgrade.

Limitations of this patch: As long as the monitor cluster does not
split brain this should work fine.  Also this assumes that NTP
among the ceph cluster is fairly accurate.

Change-Id: I04e227ea0713f0eaa4b4d78ad856e3f06fa7f225