Pause/resume for ceph charm
This changeset provides pause and resume actions to the ceph charm.
The pause action issues a 'ceph osd out <local_id>' for each of the
ceph osd ids that are on the unit.  The action does not stop the
ceph osd processes.

Note that if the pause-health action is NOT used on the ceph charm then the
cluster will start trying to rebalance the PGs accross the remaining OSDs.  If
the cluster might reach its 'full ratio' then this will be a breaking action.
The charm does NOT check for this eventuality.

The resume action issues a 'ceph osd in <local_id>' for each of the
local ceph osd process on the unit.

The charm 'remembers' that a pause action was issued, and if
successful, it shows a 'maintenance' workload status as a reminder.

Change-Id: Ic5b5b33e59e72e13843d874a08e3d142a1befde3